Refactored some parts common between surface and decoration
